create desktop app component and restructure to match figma
