+++ apps/cxc/components/FormHelpers.tsx
@@ -102,7 +102,7 @@ const selectItemStyles: Record<FormFieldVariant, string> = {
     "text-slate-200 focus:bg-slate-600 focus:text-white hover:bg-slate-600 hover:text-white transition-colors",
   auth: "text-slate-200 focus:text-white hover:!bg-slate-600/50 hover:text-white rounded-sm px-3 py-3.5 hover:bg-grey4 xl:px-4 xl:py-4 text-base",
   application:
-    "!text-gray-200 focus:!bg-zinc-700 focus:!text-white hover:!bg-zinc-700 hover:!text-white transition-colors",
+    "!text-gray-200 focus:!bg-zinc-700 focus:!text-white hover:!bg-zinc-700 hover:!text-white transition-colors p-2 text-base",
 };
 
 const textareaStyles: Record<FormFieldVariant, string> = {
