+++ packages/ui/src/index.ts
@@ -174,4 +174,4 @@ export {
 
 export { ArrowLeft, ArrowRight, Clock, MapPin, Link } from "lucide-react";
 
-export { ArrowRightIcon } from "@phosphor-icons/react";
+export { ArrowRightIcon, ArrowLeftIcon } from "@phosphor-icons/react";
