+++ packages/ui/src/styles/globals.css
@@ -43,6 +43,7 @@
 
   /* cxc */
   --cxc-input-bg: oklch(0.2478 0 0);
+  --cxc-app-bg: oklch(0.2308 0 0);
 
   /* web */
   --nav-hover-blue: oklch(61.53% 0.16853 255.741);
@@ -87,6 +88,7 @@
 
   /* cxc */
   --cxc-input-bg: oklch(0.2478 0 0);
+  --cxc-app-bg: oklch(0.2308 0 0);
 
   /* web */
   --nav-hover-blue: oklch(61.53% 0.16853 255.741);
@@ -140,6 +142,7 @@ html {
 
   /* cxc */
   --color-cxc-input-bg: var(--cxc-input-bg);
+  --color-cxc-app-bg: var(--cxc-app-bg);
 
   /* web */
   --color-nav-hover-blue: var(--nav-hover-blue);
